Finish logger implementation
Logging is now fully GO!
- We have the logger integrated with FastAPI middleware for debuging routes.
- Logger is also stand-alone callable so we can log from ANYWHERE in the app.
- Logger format shows which logger the call is coming from (file and line number).
- Logger supports std stream, file, and papertrail!

Very cool!
