+++ requirements.txt
@@ -1,10 +1,10 @@
 rich
 fastapi==0.109.2
-uvicorn==0.26.0
+uvicorn==0.27.1
 pydantic==2.6.1
 jinja2==3.1.3
 pymongo[srv]>=4.6.1
-python-multipart>=0.0.8
+python-multipart>=0.0.9
 python-jose[cryptography]>=3.3.0
 passlib[bcrypt]>=1.7.4
 httpx>=0.26.0
